requirements.txt
django>=4.0.0
djangorestframework
pyyaml
requests
django-cors-headers